Bump version to 1.4.3 for PyPI release
Summary:

This pull request bumps the `basicsr` version to 1.4.3 to trigger a new release on PyPI. The new release will include an important fix for a `torchvision` import error that affects users relying on the latest PyPI version.

Issue:

The current version of `basicsr` (1.4.2) on PyPI is missing a fix for a `ModuleNotFoundError` related to importing the `rgb_to_grayscale` function from `torchvision`.  This fix is already present in the main branch of the repository, but it hasn't been released to PyPI yet. (e.g. GFPGAN / ESRGAN in their colab notebook)

Solution:

By bumping the version number in the `VERSION` file, this pull request will trigger a new release on PyPI, making the fix available to all users who install `basicsr` through `pip`.

Testing:

I have confirmed that this fix works correctly in my local environment.